Change edit tools menu label
Previously the label for editing and tools aside menu was taken from
mediawiki message 'vies', whoch was a bit confusing (in English it's
"Views" by default, which quite does not cover the purpose and content
of this menu dropdown).

That's why I decided to move that to skin-specific i18n message.

Signed-off-by: Petr Kajzar <petr.kajzar@lf1.cuni.cz>