Change license file but preserve CC0
The LICENSE file was misleading as the repo contains bootstrap and
popper files under MIT license. Therefore I decided to change the text
of the COPYING file to cover these includes. The CC0 license text will
be included in a separate text file.

I could connect bootstrap and popper from some CDNs to not be included
inside this repo, but it would raise some problems:

* CDN urls can change or be unavailable at some time,
* MediaWiki users are likely to be using this software behind some kind
  of firewall or at local machines withou internet connection - this
  would break that.

That's why I have to be specific about the MIT license of the included
libraries.

Signed-off-by: Petr Kajzar <petr.kajzar@lf1.cuni.cz>