Fix Atom feed icon position in the dropdown menu
The feed icon in the TOOLBOX menu was displayed partially off the
margin. This fixes the position and sets it 6px from the left margin
horizontally and centered vertically.

a.feedlink also contains left padding, which is needed so that the text
is aligned in one line with the other menu items.