Use echo notifications popups on smaller screens
For screens <= 650 px it's very painful to open the notifications popup,
it does not display properly and breaks the page. That's why this has
been disabled and hidden by CSS for widths <= 650 px.

Now I would like to try another approach: let's show the echo
notifications icons (badges), but disable the onclick event and let
MediaWiki redirect to Special:Notifications page instead, which is quite
default link under the hood of the badges.