Change source of navbar's sitename
Until now, the sitename for the navbar was taken from the
MediaWiki:Sitetitle page. It is a configuration text available for
admins, however, some wikis can have more admins than needed and this
could be a vulnerable place.

Since MediaWiki:Sitetitle is usually "{{SITENAME}}", this change should
not make any problem.

It is recommended to change $wgSitename in the LocalSettings.php when
changing the sitename, not MediaWiki:Sitetitle. This is the way
other skins do.