Specify the default font size
This addresses issue #33 that caused an internal error when setting
unappropriate $wgDefaultUserOptions['medik-font'].

However, I am keeping this open as I plan to make this option more
flexible for both the users and admins. Probably this could be done with
<input type="range"> or something like that.

This commit allows admins to set literally anything to
$wgDefaultUserOptions['medik-font'], but the user will see "90%" as the
default in their user settings. And that's something I would like to fix
later.